Ajustando expires date

+++ SwiftUI/SwiftUITest/HabitPlus/HabitPlus/Common/Model/UserAuth.swift
@@ -4,7 +4,7 @@ struct UserAuth: Codable {
     
     var idToken: String
     var refreshToken: String
-    var expires: Int = 0
+    var expires: Double = 0.0
     var tokeyType: String
 
     enum CodingKeys: String, CodingKey {
